+++ Website/package.json
@@ -60,7 +60,7 @@
     "markdown-to-jsx": "^7.4.0",
     "marked-react": "^1.1.2",
     "material-icons": "^1.10.8",
-    "material-ui-confirm": "^3.0.9",
+    "material-ui-confirm": "^3.0.10",
     "monaco-editor": "^0.45.0",
     "monaco-editor-core": "^0.43.0",
     "monaco-languageclient": "^6.5.0",
