+++ Website/src/hooks/useModulesFilter.tsx
@@ -20,31 +20,31 @@ export const filters = [
     name: "No filter",
     icon: UpdateDisabledIcon,
     value: "none",
-    allowedIds: ["explore", "local"],
+    allowedIds: ["explore", "local", "update"],
   },
   {
-    name: "By date (newest)",
+    name: "Newest date first",
     icon: CalendarMonthIcon,
     value: "date_newest",
-    allowedIds: ["explore", "local"],
+    allowedIds: ["explore", "local", "update"],
   },
   {
-    name: "By date (oldest)",
+    name: "Oldest date first",
     icon: CalendarMonthIcon,
     value: "date_oldest",
-    allowedIds: ["explore", "local"],
+    allowedIds: ["explore", "local", "update"],
   },
   {
-    name: "Alphabetically",
+    name: "Name (A to Z)",
     icon: AbcIcon,
     value: "alphabetically",
-    allowedIds: ["explore", "local"],
+    allowedIds: ["explore", "local", "update"],
   },
   {
-    name: "Alphabetically (reverse)",
+    name: "Name (Z to A)",
     icon: AbcIcon,
     value: "alphabetically_reverse",
-    allowedIds: ["explore", "local"],
+    allowedIds: ["explore", "local", "update"],
   },
   // {
   //   name: "Most stars",
