+++ Website/src/locales/de.json
@@ -1,4 +1,6 @@
 {
+  "changelog": "Changelog",
+  "new": "Neu",
   "search": "Suche",
   "updates": "Updates",
   "versions": "Versionen",
