+++ Website/src/locales/en.json
@@ -1,4 +1,6 @@
 {
+  "changelog": "Changelog",
+  "new": "New",
   "search": "Search",
   "updates": "Updates",
   "versions": "Versions",
